authorization: Add OAuth support
If a JIRA username or password are provided, both are necessary
and Basic Authentication is used. If neither are provided, OAuth
is used instead. If an access token and secret are stored in the
config file, they're used and a client is created. Otherwise,
creating the JIRA client initiates an OAuth handshake. In both
cases, the consumer secret and private key filename must be stored
in the configuration file.