deployment: Add a dummy config file to enable saving
viper.ConfigFileUsed() is just a getter for the file that's set
if we call viper.SetConfigFile(). If no config file is provided,
because we're letting it use the default file, then
viper.ConfigFileUsed() returns an empty string, which causes our
SaveConfig() function to fail. This is a stopgap measure to fix
our deployment until the problem can be properly solved.